core: Use local memory events for OOMPolicy= where available
In 5.2, we (as predicted in the comment) switched to report memory
events recursively, in line with the rest of how cgroup v2 operates.
memory.events.local is introduced in 5.3+ to allow people with this
particular use case a more efficient way to avoid spurious events, so
use that where possible.