bpf: pid1: Pin reference to BPF programs for post-coldplug
During `daemon-reload` we detach and reattach all BPF programs attached
to cgroups. This, however, poses a real practical problem for
DevicePolicy (and some other settings using BPF): it presents a period
of time where the old device filtering BPF program has been unloaded,
but the new one has not been loaded yet.

Since the filtering is at open() time, on Facebook servers, we've seen
that there's a non-trivial period where applications inside that
ostensibly filtered cgroup can grab any device, and often do so, and
then retain access to that device even after the reload is over. Due to
the file continuing to be available after the initial open(), this issue
is particularly visible for DevicePolicy={strict,closed}, but also
applies to other BPF programs we install.

These BPF programs are fundamentally attached to a cgroup lifetime, not
our unit lifetime, so it's enough to pin these programs by taking a
reference to affected BPF programs before daemon-reload. 2048 programs
is a number plucked out of thin air: small enough to be safe too store
on the stack, but probably large enough to store all BPF programs in any
cromulent configuration.

That means that the BPF program lifecycle during daemon-reload changes
from this:

    manager_clear_jobs_and_units
                 │
          ╔══════╪═════════╤═══════╗
          ║ prog │ no prog │ prog' ║
          ╚══════╧═════════╪═══════╝
                           │
                    manager_coldplug

to this:

    manager_clear_jobs_and_units            unpin_all_cgroup_bpf_programs
                 │                                       │
          ╔══════╪═══════════════╤═══════════════════════╪═══════╗
          ║ prog │ prog (orphan) │ prog (orphan) + prog' │ prog' ║
          ╚══════╧═══════════════╪═══════════════════════╧═══════╝
                                 │
                          manager_coldplug

None of the BPF programs we install require exclusive access, so having
multiple instances of them running at the same time is fine.